Extend classes

+++ src/BTP.java
@@ -38,12 +38,12 @@ public class BTP   {
 
             if (firstNode) {
                 printSpace(left);
-                if (current.getElement() != 0)
+                if ((int)current.getElement() != 0)
                     System.out.print(current.getElement());
                 dist = 2 * left;
                 firstNode = false;
 
-                if (current.getElement() != 0) {
+                if ((int)current.getElement() != 0) {
                     if (current.getLeftChild() != null) {
                         children.add(current.getLeftChild());
                         moreNodes = true;
@@ -59,7 +59,7 @@ public class BTP   {
                     children.add(dummy);
                 }
             } else {
-                if (current.getElement() != 0) {
+                if ((int)current.getElement() != 0) {
                     printSpace(dist - 1);
                     System.out.print(current.getElement());
                     if (current.getLeftChild() != null) {
